# src/main/resources/openapi/weatherblanket.yaml
openapi: 3.0.3
info:
  title: Temperature Blanket API
  description: Personal API for temperature and location recording and querying. Secured using JWT.
  version: 1.0.0

security:
  - bearerAuth: []

servers:
  - url: http://157.173.123.125:8080
    description: Prod
  - url: http://localhost:8080
    description: Local

paths:
  /login:
    post:
      summary: Authenticate to obtain a JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - password
              properties:
                user:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid Credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /temperature-records:
    get:
      summary: Retrieve temperature records
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: year
          in: query
          required: false
          schema:
            type: string
        - name: month
          in: query
          required: false
          schema:
            type: string
        - name: order
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemperatureRecordResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /temperature-records/{date}/done:
    put:
      summary: Mark the selected day as Done
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Status successfully changed to Done
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /location:
    get:
      summary: Get current location
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: Update location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /location/reset:
    put:
      summary: Reset to base location
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LocationRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double

    LoginResponse:
      type: object
      required:
        - access_token
      properties:
        access_token:
          type: string

    TemperatureRecordResponse:
      type: object
      required:
        - date
        - time
        - temperature
        - color
        - yarn_color
        - status
      properties:
        date:
          type: string
        time:
          type: string
        temperature:
          type: number
          format: double
        color:
          type: string
        yarn_color:
          type: string
        status:
          type: string

    LocationResponse:
      type: object
      required:
        - latitude
        - longitude
        - city
        - state
        - country
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        city:
          type: string
        state:
          type: string
        country:
          type: string

    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string

  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Token ausente, expirado ou inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalError:
      description: Erro interno no servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'